requirements.txt
# Django Core
Django>=5.0,<5.1
python-decouple>=3.8

# Database
psycopg2-binary>=2.9.9  # PostgreSQL
# mysqlclient>=2.2.0    # MySQL (uncomment if needed)

# REST API (if needed)
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0

# Storage & Files
Pillow>=10.1.0
django-storages>=1.14.2

# Caching & Performance
redis>=5.0.1
django-redis>=5.4.0
celery>=5.3.4

# Reporting
reportlab>=4.0.7
openpyxl>=3.1.2
pandas>=2.1.4

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3

# Development
django-debug-toolbar>=4.2.0
black>=23.12.0
flake8>=6.1.0
pytest-django>=4.7.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
Django>=4.2
pymysql>=1.1.0
python-dotenv>=1.0.0
cryptography>=41.0.0